Fix duplicated events
